Test converting balance to UINT16, rather than unsigned int
Issue #119
-e 
Original/build commit ID: ba875fa0e144d2eb369cde1f6de5cc53b3d51889

+++ balance.h
@@ -7,6 +7,6 @@
 
 #include "gb.h"
 
-BOOLEAN has_money(unsigned int amount_h, unsigned int amount_l);
-BOOLEAN remove_money(unsigned int amount_h, unsigned int amount_l);
-void add_money(unsigned int amount_h, unsigned int amount_l);
+BOOLEAN has_money(UINT16 amount_h, UINT16 amount_l);
+BOOLEAN remove_money(UINT16 amount_h, UINT16 amount_l);
+void add_money(UINT16 amount_h, UINT16 amount_l);
